+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Fixed
+
+- `be`: Update regex to account for new VAT numbers starting with 1.
+
 ## [v0.302.1] - 2025-10-31
 
 ### Fixed